requirements.txt
# ── Core Web Framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# ── Environment ───────────────────────────────────────────────────
python-dotenv==1.0.1

# ── AI / ML Core ──────────────────────────────────────────────────
openai-whisper==20231117        # Speech transcription
openai==1.30.0                  # LLM (GPT-4o-mini)

# ── Audio Analysis ────────────────────────────────────────────────
librosa==0.10.1                 # Audio feature extraction
soundfile==0.12.1               # Audio I/O

# ── Video Processing ─────────────────────────────────────────────
moviepy==1.0.3                  # Video editing
opencv-python-headless==4.9.0.80 # Computer vision (no GUI deps)

# ── Face Detection ────────────────────────────────────────────────
mediapipe==0.10.14              # Face tracking for smart crop

# ── Google Gemini (alternative LLM) ─────────────────────────────
google-generativeai==0.7.2

# ── Data / Utilities ─────────────────────────────────────────────
numpy==1.26.4
pydantic==2.7.1
httpx==0.27.0

# ── Optional: faster Whisper ─────────────────────────────────────
# faster-whisper==1.0.1       # Drop-in replacement, 4x faster
